# public/swagger.json
{
	"openapi": "3.0.0",
	"components": {
		"examples": {},
		"headers": {},
		"parameters": {},
		"requestBodies": {},
		"responses": {},
		"schemas": {
			"User": {
				"properties": {
					"userId": {
						"type": "string"
					},
					"firstName": {
						"type": "string"
					},
					"lastName": {
						"type": "string"
					},
					"fullName": {
						"type": "string"
					},
					"email": {
						"type": "string"
					},
					"password": {
						"type": "string"
					},
					"phoneNumber": {
						"type": "string"
					},
					"address": {
						"type": "string"
					},
					"isActive": {
						"type": "boolean"
					},
					"createdDate": {
						"type": "string",
						"format": "date-time"
					},
					"updatedDate": {
						"type": "string",
						"format": "date-time"
					}
				},
				"required": [
					"userId",
					"firstName",
					"lastName",
					"fullName",
					"email",
					"password",
					"phoneNumber",
					"address",
					"isActive",
					"createdDate",
					"updatedDate"
				],
				"type": "object",
				"additionalProperties": false
			},
			"LoginUser": {
				"properties": {
					"userId": {
						"type": "string"
					},
					"firstName": {
						"type": "string"
					},
					"lastName": {
						"type": "string"
					},
					"fullName": {
						"type": "string"
					},
					"email": {
						"type": "string"
					},
					"phoneNumber": {
						"type": "string"
					},
					"address": {
						"type": "string"
					},
					"token": {
						"type": "string"
					},
					"isActive": {
						"type": "boolean"
					},
					"createdDate": {
						"type": "string",
						"format": "date-time"
					},
					"updatedDate": {
						"type": "string",
						"format": "date-time"
					}
				},
				"required": [
					"userId",
					"firstName",
					"lastName",
					"fullName",
					"email",
					"phoneNumber",
					"address",
					"isActive",
					"createdDate",
					"updatedDate"
				],
				"type": "object",
				"additionalProperties": false
			},
			"CreateUser": {
				"properties": {
					"firstName": {
						"type": "string"
					},
					"lastName": {
						"type": "string"
					},
					"fullName": {
						"type": "string"
					},
					"email": {
						"type": "string"
					},
					"password": {
						"type": "string"
					},
					"phoneNumber": {
						"type": "string"
					},
					"address": {
						"type": "string"
					}
				},
				"required": [
					"firstName",
					"lastName",
					"fullName",
					"email",
					"password",
					"phoneNumber",
					"address"
				],
				"type": "object",
				"additionalProperties": false
			}
		},
		"securitySchemes": {
			"bearerAuth": {
				"type": "http",
				"scheme": "bearer",
				"bearerFormat": "JWT"
			}
		}
	},
	"info": {
		"title": "login_portal_api",
		"version": "1.0.0",
		"description": "login portal to authorise the user",
		"license": {
			"name": "ISC"
		},
		"contact": {}
	},
	"paths": {
		"/api/users/allUsers": {
			"get": {
				"operationId": "GetAllUsers",
				"responses": {
					"200": {
						"description": "Ok",
						"content": {
							"application/json": {
								"schema": {
									"items": {
										"$ref": "#/components/schemas/User"
									},
									"type": "array"
								}
							}
						}
					}
				},
				"tags": [
					"Users"
				],
				"security": [
					{
						"bearerAuth": []
					}
				],
				"parameters": []
			}
		},
		"/api/auth/register": {
			"post": {
				"operationId": "CreateUser",
				"responses": {
					"200": {
						"description": "Ok",
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/LoginUser"
								}
							}
						}
					}
				},
				"tags": [
					"Auth"
				],
				"security": [],
				"parameters": [],
				"requestBody": {
					"required": true,
					"content": {
						"application/json": {
							"schema": {
								"$ref": "#/components/schemas/CreateUser"
							}
						}
					}
				}
			}
		},
		"/api/auth/login": {
			"post": {
				"operationId": "LoginUser",
				"responses": {
					"200": {
						"description": "Ok",
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/LoginUser"
								}
							}
						}
					}
				},
				"tags": [
					"Auth"
				],
				"security": [],
				"parameters": [],
				"requestBody": {
					"required": true,
					"content": {
						"application/json": {
							"schema": {
								"properties": {
									"password": {
										"type": "string"
									},
									"email": {
										"type": "string"
									}
								},
								"required": [
									"password",
									"email"
								],
								"type": "object"
							}
						}
					}
				}
			}
		}
	},
	"servers": [
		{
			"url": "/"
		}
	]
}